rename DBs and use straxen entrypoints
